+++ NEWS.md
@@ -74,3 +74,9 @@ via the `decode` parameter.
 `causal_words_pred()`, and `causal_pred_mats()` instead.
 - Deprecated `masked_tokens_tbl()` and `masked_lp()`. Use 
 `masked_tokens_pred_tbl()` and `masked_targets_pred()` instead.
+
+
+# pangoling 0.0.0.9011
+
+- Added `word_n` argument in `causal_words_pred()` to indicate word order of teh texts.
+- Allows for models with larger vocabulary than tokenizer.